Updated AutoConf examples

+++ BashExamples/00_hello_world/main.cpp
@@ -3,8 +3,8 @@
 
 int main() {
   std::vector<int> a = {1, 2, 3};
-  for (int i = 0; i < a.size() - 5; ++i) {
-	std::cout << a[i] << " ";
+  for (int i = 0; i < static_cast<int>(a.size()) - 5; ++i) {
+	  std::cout << a[i] << " ";
   }
   std::cout << std::endl;
   std::cout << "Hello world!" << std::endl;
